% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseArguments.R
\name{ParseArguments}
\alias{ParseArguments}
\title{ParseArguments}
\usage{
ParseArguments(accepted.switches, command.args = commandArgs(trailingOnly =
  TRUE))
}
\arguments{
\item{accepted.switches}{Named list of switches. The names of
accepted.switches will be used as the names of the parsed arguments.}

\item{command.args}{Character vector of command arguments. Defaults to
\code{commandArgs(trailingOnly = TRUE)}, \emph{i.e.} the trailing
command-line arguments.}
}
\value{
Returns a named list of command-line arguments. The names are taken
  from \code{names(accepted.switches)}.
}
\description{
Parse command-line arguments
}
\examples{

ParseArguments(list(output = "-o", input = "-i"),
  c("-i", "output/covar_analysis/post_recal_data.table",
    "-i", "output/covar_analysis/recal_data.table",
    "-o", "output/covar_analysis/recalibration_plots.pdf")))

}

